Added Files "Screen"
Files .h .c added for screen.

Code compiling without warning or mistakes.
